✨ feat(app): add Expo Dev Client and update dependencies
This commit introduces the Expo Dev Client to the project and updates several dependencies.

The Expo Dev Client allows for faster development builds and improved debugging capabilities. It is added as a plugin in `app.json`.

The following dependencies are updated:

- `@expo/metro-runtime`
- `@expo/vector-icons`
- `@react-native-async-storage/async-storage`
- `@react-native-clipboard/clipboard`
- `@react-native-community/slider`
- `@react-navigation/bottom-tabs`
- `@react-navigation/drawer`
- `@react-navigation/native`
- `expo`
- `expo-av`
- `expo-file-system`
- `expo-font`
- `expo-linking`
- `expo-router`
- `expo-status-bar`
- `nativewind`
- `prettier-plugin-tailwindcss`
- `react`
- `react-dom`
- `react-native`
- `react-native-gesture-handler`
- `react-native-reanimated`
- `react-native-safe-area-context`
- `react-native-screens`
- `react-native-vector-icons`
- `react-native-web`
- `tailwindcss`

The `sura-list.js` file is also updated to improve code readability.